Add storybook to react app
Prior to this, you need to install the storybook cli app globally, with:

    yarn global add @storybook/cli

Then, starting from your bog-standard React app, in this case generated by `create-react-app`, we add storybook to it with the command:

    getstorybook

This creates the `.storybook` directory for configuration data, and the `src/stories/` folder, with a sample story in `src/stories/index.js`.

At this point, you can run the storybook server with:

    yarn run storybook

This starts a server running at `http://localhost:9009` which you can open in your browser.