Add a component and a story
In this step, we added a horizontal form input field component, as well as some styling components based on `styled-component`.

The form group component is not doing anything all that interesting, but we can see how our formatting choices play out without needing to bring up the full application.

The story "consumes" the component, passing in the required properties for a 4 field form.