Using a Consumer
Using a Consumer can dry up the story, so your stories only have to pass information unique to each story, and not any of the boilerplate or other things necessary to encapsulate the Component.

LoginForm now gets a prop containing any pre-existing account information, such as for an edit operation. The story file has two stories, one when the account prop is not passed, and one with a passed account prop. The form has to pass any information provided in the account prop into it's state.